Edited Server.js file
